feat: create hook to use calculator
